# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Sistema AXXESS API",
        "description": "API REST para el Sistema de Gestión AXXESS - GLAZ GROUP S.A.S.",
        "contact": {
            "name": "GLAZ GROUP S.A.S.",
            "email": "soporte@glazgroup.com"
        },
        "license": {
            "name": "Confidencial",
            "url": "https://glazgroup.com/licencia"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "Servidor de Desarrollo"
        }
    ],
    "paths": {
...
        "/api/dashboard/general": {
            "get": {
                "tags": [
                    "Dashboard"
                ],
                "summary": "Estadísticas generales del sistema",
                "description": "Retorna métricas globales del sistema AXXESS",
                "operationId": "a3a063e1318b5b747a478233f4fb0f53",
                "responses": {
                    "200": {
                        "description": "Estadísticas generales",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "productos_activos": {
                                                    "type": "integer",
                                                    "example": 7
                                                },
                                                "productos_stock_bajo": {
                                                    "type": "integer",
                                                    "example": 3
                                                },
                                                "pacientes_activos": {
                                                    "type": "integer",
                                                    "example": 3
                                                },
                                                "clientes_activos": {
                                                    "type": "integer",
                                                    "example": 3
                                                },
                                                "pedidos_pendientes": {
                                                    "type": "integer",
                                                    "example": 0
                                                },
                                                "pedidos_hoy": {
                                                    "type": "integer",
                                                    "example": 0
                                                },
                                                "ventas_mes": {
                                                    "type": "number",
                                                    "example": 0
                                                },
                                                "ventas_hoy": {
                                                    "type": "number",
                                                    "example": 0
                                                },
                                                "lotes_proximos_vencer": {
                                                    "type": "integer",
                                                    "example": 0
                                                },
                                                "alertas_no_leidas": {
                                                    "type": "integer",
                                                    "example": 0
                                                },
                                                "alertas_criticas": {
                                                    "type": "integer",
                                                    "example": 0
                                                },
                                                "cuentas_vencidas": {
                                                    "type": "integer",
                                                    "example": 0
                                                },
                                                "valor_cuentas_por_cobrar": {
                                                    "type": "number",
                                                    "example": 0
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/dashboard/ventas": {
            "get": {
                "tags": [
                    "Dashboard"
                ],
                "summary": "Estadísticas de ventas",
                "description": "Retorna métricas de ventas por período",
                "operationId": "81be516a334affd69c89d29d9f03818d",
                "parameters": [
                    {
                        "name": "periodo",
                        "in": "query",
                        "description": "Período de consulta",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "dia",
                                "semana",
                                "mes",
                                "año"
                            ],
                            "example": "mes"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Estadísticas de ventas"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/dashboard/inventario": {
            "get": {
                "tags": [
                    "Dashboard"
                ],
                "summary": "Estadísticas de inventario",
                "description": "Retorna métricas del inventario actual",
                "operationId": "22a87a32a8ad79ad46eb606e560cc45f",
                "responses": {
                    "200": {
                        "description": "Estadísticas de inventario"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/dashboard/cartera": {
            "get": {
                "tags": [
                    "Dashboard"
                ],
                "summary": "Estadísticas de cartera",
                "description": "Retorna métricas de cuentas por cobrar",
                "operationId": "cab5c9f251a8dc5579e00f49f48bb28b",
                "responses": {
                    "200": {
                        "description": "Estadísticas de cartera"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/dashboard/therapies": {
            "get": {
                "tags": [
                    "Dashboard"
                ],
                "summary": "Estadísticas módulo THERAPIES",
                "description": "Retorna métricas del módulo THERAPIES",
                "operationId": "ddcf45c7f75ee3d3f19bb7c0216308a2",
                "responses": {
                    "200": {
                        "description": "Estadísticas THERAPIES"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/dashboard/market": {
            "get": {
                "tags": [
                    "Dashboard"
                ],
                "summary": "Estadísticas módulo MARKET",
                "description": "Retorna métricas del módulo MARKET",
                "operationId": "d7a6dd6ea9792650f39c085eb5b1fd2b",
                "responses": {
                    "200": {
                        "description": "Estadísticas MARKET"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/dashboard/alertas": {
            "get": {
                "tags": [
                    "Dashboard"
                ],
                "summary": "Resumen de alertas",
                "description": "Retorna métricas del sistema de alertas",
                "operationId": "cb21a00af89bb8a5290a942b80d81a43",
                "responses": {
                    "200": {
                        "description": "Resumen de alertas"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Ingrese el token de autenticación Sanctum",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}